refactor: code refs: https://github.com/RonasIT/laravel-telescope-extension/issues/5
